Updated GetOBCValues1: Made code loop if get iphonesubinfo is empty.
Does this by checking the length of the IMEI and if it is a number. Will wait
and loop if correct IMEI is not found. This change is necessary because
sometimes IMEI and Serial Number was not being written to the file and I
suspect it is because if iphonesubifo was called too soon then if would
not contain the IMEI. Restructured code as well