REL: v1.2.38.1
This commit fixed a bug with the audio test that Eyal reported. Now the
audio test turns of the speakers multiple times to ensure that you only
hear one speaker at a time.

The main changes to this release involve the bug fix to the audio test
and fixing a date issue with the database. Also changes to the Android
Studio project .gitignore file.