Test Version v1.2.39-dev-2:
Both the database and individual result files use the same local
datetime from sqlite. Result file contains "1" for pass or "0" for fail
and "NA" for not run tests. Added single quotes to imei, serial, board
serial to avoid the problem where excel auto formats the value as an
integer. Any tests run that end with "_ud" will have that part of the
name removed when the results are saved to the database.